ConnectionsRetrievalActor-s common logic to abstract class
