+++ gateway/service/src/main/java/org/eclipse/ditto/gateway/service/endpoints/actors/AbstractConnectionsRetrievalActor.java
@@ -13,21 +13,45 @@
 
 package org.eclipse.ditto.gateway.service.endpoints.actors;
 
-import java.time.Duration;
+import static org.eclipse.ditto.base.model.common.ConditionChecker.checkNotNull;
 
+import java.time.Duration;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
+import java.util.Optional;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.CompletionStage;
+
+import org.eclipse.ditto.base.model.exceptions.DittoRuntimeException;
+import org.eclipse.ditto.base.model.headers.DittoHeaders;
+import org.eclipse.ditto.base.model.headers.WithDittoHeaders;
+import org.eclipse.ditto.base.model.signals.commands.Command;
+import org.eclipse.ditto.base.model.signals.commands.ErrorResponse;
+import org.eclipse.ditto.connectivity.model.ConnectionId;
 import org.eclipse.ditto.connectivity.model.ConnectivityInternalErrorException;
 import org.eclipse.ditto.connectivity.model.signals.commands.ConnectivityErrorResponse;
+import org.eclipse.ditto.connectivity.model.signals.commands.exceptions.ConnectionsAmountIllegalException;
+import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveAllConnectionIds;
+import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveAllConnectionIdsResponse;
+import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveConnection;
+import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveConnectionResponse;
 import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveConnections;
+import org.eclipse.ditto.connectivity.model.signals.commands.query.RetrieveConnectionsResponse;
 import org.eclipse.ditto.gateway.service.util.config.DittoGatewayConfig;
 import org.eclipse.ditto.gateway.service.util.config.endpoints.CommandConfig;
 import org.eclipse.ditto.internal.utils.akka.logging.DittoLoggerFactory;
 import org.eclipse.ditto.internal.utils.akka.logging.ThreadSafeDittoLogger;
 import org.eclipse.ditto.internal.utils.config.DefaultScopedConfig;
+import org.eclipse.ditto.json.JsonArray;
+import org.eclipse.ditto.json.JsonCollectors;
 
 import akka.actor.AbstractActor;
 import akka.actor.ActorRef;
 import akka.actor.ReceiveTimeout;
 import akka.japi.pf.ReceiveBuilder;
+import akka.pattern.Patterns;
 
 
 /**
@@ -65,13 +89,55 @@ public abstract class AbstractConnectionsRetrievalActor extends AbstractActor {
     }
 
     protected abstract void retrieveConnections(final RetrieveConnections retrieveConnections);
-    protected abstract Receive commandResponseAwaitingBehaviour();
+
+    protected void retrieveAllConnectionsIds(final RetrieveConnections retrieveConnections, final ActorRef sender) {
+        this.edgeCommandForwarder.tell(RetrieveAllConnectionIds.of(retrieveConnections.getDittoHeaders()), sender);
+    }
+
+    protected void retrieveConnectionsById(RetrieveAllConnectionIdsResponse msg) {
+        final var connectionIds = msg.getAllConnectionIds();
+        if (initialCommand.getIdsOnly()) {
+            RetrieveConnectionsResponse response = RetrieveConnectionsResponse
+                    .of(JsonArray.of(connectionIds), initialCommand.getDittoHeaders());
+            sender.tell(response, getSelf());
+            stop();
+        } else {
+            if (connectionIds.size() > connectionsRetrieveLimit) {
+                sender.tell(ConnectionsAmountIllegalException.newBuilder(connectionsRetrieveLimit)
+                        .dittoHeaders(initialCommand.getDittoHeaders())
+                        .build(), getSelf());
+                stop();
+            } else {
+                retrieveConnections(connectionIds
+                        .stream()
+                        .map(ConnectionId::of)
+                        .toList(), initialCommand.getDittoHeaders())
+                        .thenAccept(retrieveConnectionsResponse -> {
+                            sender.tell(retrieveConnectionsResponse, getSelf());
+                            stop();
+                        })
+                        .exceptionally(t -> {
+                            ConnectivityInternalErrorException exception =
+                                    ConnectivityInternalErrorException.newBuilder()
+                                            .dittoHeaders(initialCommand.getDittoHeaders())
+                                            .cause(t)
+                                            .build();
+                            ConnectivityErrorResponse response =
+                                    ConnectivityErrorResponse.of(exception, initialCommand.getDittoHeaders());
+                            sender.tell(response, getSelf());
+                            stop();
+                            return null;
+                        });
+            }
+        }
+    }
 
     private Receive responseAwaitingBehavior(){
-        return commandResponseAwaitingBehaviour().orElse(ReceiveBuilder.create()
+        return ReceiveBuilder.create()
+                .match(RetrieveAllConnectionIdsResponse.class, this::retrieveConnectionsById)
                 .match(ReceiveTimeout.class, timeout -> handleTimeout())
                 .matchAny(msg -> logger.warn("Unknown message: <{}>", msg))
-                .build());
+                .build();
     }
 
     private void handleRetrieveConnections(RetrieveConnections retrieveConnections) {
@@ -80,6 +146,103 @@ public abstract class AbstractConnectionsRetrievalActor extends AbstractActor {
         retrieveConnections(retrieveConnections);
     }
 
+    private CompletionStage<RetrieveConnectionsResponse> retrieveConnections(
+            final Collection<ConnectionId> connectionIds, final DittoHeaders dittoHeaders) {
+
+        checkNotNull(connectionIds, "connectionIds");
+        checkNotNull(dittoHeaders, "dittoHeaders");
+
+        final List<CompletableFuture<RetrieveConnectionResponse>> completableFutures = connectionIds.parallelStream()
+                .map(connectionId -> retrieveConnection(RetrieveConnection.of(connectionId, dittoHeaders)))
+                .map(CompletionStage::toCompletableFuture)
+                .toList();
+
+        final List<ConnectionId> connectionIdList = new ArrayList<>(connectionIds);
+
+        // this comparator is used to support returning the connections in the same order they were requested
+        final Comparator<RetrieveConnectionResponse> sorter = (r1, r2) -> {
+            final ConnectionId r1ConnectionId = r1.getEntityId();
+            final ConnectionId r2ConnectionId = r2.getEntityId();
+            return Integer.compare(connectionIdList.indexOf(r1ConnectionId), connectionIdList.indexOf(r2ConnectionId));
+        };
+
+        return CompletableFuture.allOf(completableFutures.toArray(new CompletableFuture[0]))
+                .thenApply(ignored -> completableFutures.stream()
+                        .map(CompletableFuture::join)
+                        .toList())
+                .thenApply(retrieveConnectionResponses -> {
+                    final JsonArray connections = retrieveConnectionResponses.stream()
+                            .sorted(sorter)
+                            .map(RetrieveConnectionResponse::getEntity)
+                            .collect(JsonCollectors.valuesToArray());
+
+                    return RetrieveConnectionsResponse.of(connections, dittoHeaders);
+                });
+    }
+
+    private CompletionStage<RetrieveConnectionResponse> retrieveConnection(
+            final RetrieveConnection retrieveConnection) {
+
+        return askConnectivity(retrieveConnection);
+    }
+
+    private CompletionStage<RetrieveConnectionResponse> askConnectivity(
+            final RetrieveConnection command) {
+
+        logger.withCorrelationId(command).debug("Sending command <{}> to org.eclipse.ditto.connectivity.service.", command);
+        final var commandWithCorrelationId = ensureCommandHasCorrelationId(command);
+        Duration askTimeout = initialCommand.getDittoHeaders().getTimeout().orElse(defaultTimeout);
+        return Patterns.ask(edgeCommandForwarder, commandWithCorrelationId, askTimeout)
+                .thenApply(response -> {
+                    logger.withCorrelationId(command)
+                            .debug("Received response <{}> from org.eclipse.ditto.connectivity.service.", response);
+                    throwCauseIfErrorResponse(response);
+                    throwCauseIfDittoRuntimeException(response);
+                    final RetrieveConnectionResponse mappedResponse =
+                            mapToType(response, RetrieveConnectionResponse.class, command);
+                    logger.withCorrelationId(command)
+                            .info("Received response of type <{}> from org.eclipse.ditto.connectivity.service.",
+                                    mappedResponse.getType());
+                    return mappedResponse;
+                });
+    }
+
+    private static Command<?> ensureCommandHasCorrelationId(final Command<?> command) {
+        final DittoHeaders dittoHeaders = command.getDittoHeaders();
+        final Optional<String> correlationIdOptional = dittoHeaders.getCorrelationId();
+        if (correlationIdOptional.isPresent()) {
+            return command;
+        }
+        return command.setDittoHeaders(dittoHeaders.toBuilder().randomCorrelationId().build());
+    }
+
+    private static void throwCauseIfErrorResponse(final Object response) {
+        if (response instanceof ErrorResponse<?> errorResponse) {
+            throw (errorResponse).getDittoRuntimeException();
+        }
+    }
+
+    private static void throwCauseIfDittoRuntimeException(final Object response) {
+        if (response instanceof DittoRuntimeException dittoRuntimeException) {
+            throw dittoRuntimeException;
+        }
+    }
+
+    private static <T> T mapToType(final Object object, final Class<T> targetClass,
+            final WithDittoHeaders withDittoHeaders) {
+
+        final Class<?> actualClass = object.getClass();
+        if (targetClass.isAssignableFrom(actualClass)) {
+            return targetClass.cast(object);
+        }
+        final String message =
+                String.format("Expected <%s> response but got <%s>!", targetClass.getSimpleName(), actualClass);
+        final IllegalArgumentException exception = new IllegalArgumentException(message);
+        throw ConnectivityInternalErrorException.newBuilder()
+                .dittoHeaders(withDittoHeaders.getDittoHeaders())
+                .cause(exception)
+                .build();
+    }
 
 
     protected void handleTimeout() {
